feat: [ANDROAPP-5502] readapt screen management
