+++ myplugin/src/main/java/org/dhis2/mobile/myplugin/MainActivity.kt
@@ -46,89 +46,18 @@ class MainActivity : ComponentActivity() {
         enableEdgeToEdge()
         setContent {
             Dhis2androidcaptureappTheme {
-                val programs = viewModel.programList.observeAsState(listOf())
 
 //                Scaffold(modifier = Modifier.fillMaxSize()) { innerPadding ->
 //
-                MainScreen(this, viewModel, programs.value)
-//                }
-            }
-        }
-    }
-
-    @Composable
-    fun MainScreen(
-        activity: Activity,
-        viewModel: MainViewModel = viewModel(),
-        programs: List<ProgramItem>
-    ) {
-
-        Scaffold(modifier = Modifier.fillMaxSize()) {
-                innerPadding ->
-
-            val programList: List<ProgramItem> = programs
-
-            LazyColumn(
-                Modifier.padding(top = 16.dp),
-            ) {
-                items(programList) {
-                    ProgramCard(program = it.program)
-                }
-            }
-            Column(modifier = Modifier.padding(innerPadding)) {
-                CardDetail(
-                    title = "This is  new content with activity lifecycle and hopefully sdk requests!!!",
-                    additionalInfoList = listOf(
-                        AdditionalInfoItem(
-                            icon = {
-                                Icon(
-                                    imageVector = Icons.Outlined.Android,
-                                    contentDescription = "This is my amazing plugin!!!",
-                                    tint = AdditionalInfoItemColor.SUCCESS.color,
-                                )
-                            },
-                            value = "Lets Rock",
-                            color = AdditionalInfoItemColor.ERROR.color,
-                            isConstantItem = true,
-                        ),
-                    ),
-                )
+                PluginImpl().Show(this)
+//               }
             }
         }
     }
 
 
-    @Composable
-    fun ProgramCard(program: Program) {
-        ListCard(title = ListCardTitleModel(text =  program.displayName() ?: "-"),
-            additionalInfoList = getAdditionalInfoList(program),
-            onCardClick = { navigateToProgramDetailScreen(program) })
-    }
-
-    private fun navigateToProgramDetailScreen(program: Program) {
-
-    }
-
-    private fun getAdditionalInfoList(program: Program): List<AdditionalInfoItem> {
-        val detailsList: MutableList<AdditionalInfoItem> = mutableListOf()
-        program.programType()?.name.let {
-            detailsList.add(AdditionalInfoItem(key = "Type", value = it?: "-"))
-        }
-
-        program.trackedEntityType()?.name().let {
-            detailsList.add(AdditionalInfoItem(key = "Tracked Entity Type", value = it?: "-"))
-        }
-
-        program.displayEventLabel().let {
-            detailsList.add(AdditionalInfoItem(key = "Event Label", value = it?: "-"))
-        }
-
-
-
-        return detailsList
-    }
 }
 
-@HiltAndroidApp
-class CoreApplication: Application()
 
+@HiltAndroidApp
+class CoreApplication : Application()
